# Don't parse response content type if value comes from Mime